+++ src/components/Feed/Proposals.js
